requirements.txt
asgiref==3.2.3
Django==3.0.2
django-extensions==2.2.6
djangorestframework==3.11.0
pydot==1.4.1
pyparsing==2.4.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
